+++ dakara_server/playlist/tests_playlist_entry.py
@@ -138,6 +138,27 @@ class PlaylistEntryListViewListCreateAPIViewTestCase(BaseAPITestCase):
         response = self.client.post(self.url, {"song": self.song1.id})
         self.assertEqual(response.status_code, status.HTTP_403_FORBIDDEN)
 
+    @patch('playlist.views.settings')
+    def test_post_create_playlist_entry_playlist_full_forbidden(
+            self, mock_settings
+            ):
+        """
+        Test to verify playlist entry creation
+        """
+        # mock the settings
+        mock_settings.PLAYLIST_SIZE_LIMIT = 1
+
+        # Login as playlist user
+        self.authenticate(self.p_user)
+
+        # Pre assert 4 entries in database
+        # (2 in queue)
+        self.assertEqual(PlaylistEntry.objects.count(), 4)
+
+        # Post new playlist entry
+        response = self.client.post(self.url, {"song": self.song1.id})
+        self.assertEqual(response.status_code, status.HTTP_403_FORBIDDEN)
+
     def test_post_create_user_forbidden(self):
         """
         Test to verify simple user cannot create playlist entries